Cleanup, minor UI changes and a few bugfixes
